+++ go/openinference-semantic-conventions/semconv_test.go
@@ -45,6 +45,8 @@ func TestAttributeKeys(t *testing.T) {
 		{EvaluationMetadata, "evaluation.metadata"},
 
 		{LLMModelName, "llm.model_name"},
+		{LLMRequestModelName, "llm.request.model_name"},
+		{LLMResponseModelName, "llm.response.model_name"},
 		{LLMProvider, "llm.provider"},
 		{LLMSystem, "llm.system"},
 		{LLMInvocationParameters, "llm.invocation_parameters"},