+++ js/packages/openinference-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @arizeai/openinference-core
 
+## 2.5.3
+
+### Patch Changes
+
+- Updated dependencies [1fe497f]
+  - @arizeai/openinference-semantic-conventions@2.8.0
+
 ## 2.5.2
 
 ### Patch Changes